Revert "feat: update api structure"
This reverts commit 809c98396933917cf3a889fffc0f0a1fec250176.

+++ src/app/api/teachers/[id]/route.ts
@@ -1,40 +1,48 @@
 import { teacherService } from "@/applications/container";
-import { ValidationError } from "@/infrastructure/helper/error";
-import responseApiHandler from "@/infrastructure/helper/rensponse-api-handler";
-import tokenValidationApiHandler from "@/infrastructure/helper/token-validation-api-handler";
+import { AuthError, ErrorInterface, ValidationError } from "@/infrastructure/helper/error";
+import { validateToken } from "@/infrastructure/helper/validate-token";
 import { UpdateTeacher } from "@/infrastructure/interfaces/teacher";
 import { NextRequest } from "next/server";
 
 export async function PATCH(request: NextRequest, { params }: { params: Promise<{ id: string }> }) {
-    return await responseApiHandler(
-        async () => {
-            tokenValidationApiHandler(request)
+    try {
+        //token validation
+        const token = request.headers.get('authorization')
+        if (!validateToken(token)) throw new AuthError('Unauthorized')
 
-            //validate id
-            const id = Number((await params).id)
-            if (isNaN(id)) throw new ValidationError('Invalid id')
+        const id = Number((await params).id)
+        //validate id
+        if (isNaN(id)) throw new ValidationError('Invalid id')
 
-            const body = await request.json();
+        const body = await request.json();
 
-            const item = await teacherService.update(id, body as UpdateTeacher)
+        const data = await teacherService.update(id, body as UpdateTeacher)
 
-            return { successCode: 200, item }
-        }
-    )
+        return Response.json({ message: 'success', data }, { status: 200 })
+    } catch (error) {
+        const _error = error as ErrorInterface
+
+        return Response.json({ message: _error.message, errors: _error.details }, { status: _error.code })
+    }
+
 }
 
 export async function DELETE(request: NextRequest, { params }: { params: Promise<{ id: string }> }) {
-    return await responseApiHandler(
-        async () => {
-            tokenValidationApiHandler(request)
+    try {
+        //token validation
+        const token = request.headers.get('authorization')
+        if (!validateToken(token)) throw new AuthError('Unauthorized')
 
-            //validate id
-            const id = Number((await params).id)
-            if (isNaN(id)) throw new ValidationError('Invalid id')
+        const id = Number((await params).id)
+        //validate id
+        if (isNaN(id)) throw new ValidationError('Invalid id')
 
-            await teacherService.delete(id)
+        await teacherService.delete(id)
 
-            return { successCode: 200 }
-        }
-    )
+        return Response.json({ message: 'success' }, { status: 200 })
+    } catch (error) {
+        const _error = error as ErrorInterface
+
+        return Response.json({ message: _error.message, errors: _error.details }, { status: _error.code })
+    }
 }
